Move `CudaDevicePlacementMixin` to new module
